requirements.txt
botcity-framework-web>=0.9.0,<1.0
botcity-maestro-sdk>=0.3.3,<1.0

# Drivers Google Chrome
webdriver_manager

# Pandas
pandas==2.2.3
